GFM-104 - refactored address lookup to remove duplicate code
* renamed 'address' to 'address-lookup'
* removed address-inside, address-outside and no-postcode
* replaced the above with one step - 'address'
* no-postcode controller now called `address`
* added `isDomestic` to address controller locals - true if UK
* added conditional panel to address - shows postcode if domestic
* abstracted edit-postcode html to partial - used by address.html and address-lookup.html
* bugfix - unset `addresses` and apiErrorMessage from sessionModel on BT postcode and success api call
