+++ apps/common/views/partials/amend-postcode.html
@@ -1,0 +1,7 @@
+<label>
+  {{#t}}fields.postcode.label{{/t}}
+</label>
+<p>
+  <span class="postcode">{{values.postcode}}</span>
+  <span class="link"><a href="{{baseUrl}}/postcode">{{#t}}pages.address.edit{{/t}}</a></span>
+</p>
