+++ apps/gro/fields/address-lookup.js
@@ -1,0 +1,7 @@
+'use strict';
+
+module.exports = {
+  'address-lookup': {
+    className: ['address']
+  }
+};
