--- apps/gro/fields/address-outside.js
@@ -1,9 +1,0 @@
-'use strict';
-
-module.exports = {
-  'address-textarea': {
-    validate: ['required'],
-    'ignore-defaults': true,
-    formatter: ['trim', 'hyphens']
-  }
-};
