--- apps/gro/views/address-inside.html
@@ -1,43 +1,0 @@
-{{<layout}}
-  {{$step}}Step 2{{/step}}
-
-  {{$journeyHeader}}
-    {{#t}}journey.header{{/t}}
-  {{/journeyHeader}}
-
-  {{$validationSummary}}
-    {{> partials-validation-summary}}
-  {{/validationSummary}}
-
-  {{$propositionHeader}}
-    {{> partials-navigation}}
-  {{/propositionHeader}}
-
-  {{$header}}{{#t}}pages.postcode.header{{/t}}{{/header}}
-
-  {{$content}}
-
-    {{<partials-form}}
-
-      {{$form}}
-
-        {{#t}}fields.postcode.label{{/t}}
-        <p>
-          <span class="postcode">{{values.postcode}}</span>
-            {{#backLink}}
-              <span class="link">
-                <a href="{{#url}}{{backLink}}{{/url}}">{{#t}}pages.address.edit{{/t}}</a>
-              </span>
-            {{/backLink}}
-        </p>
-
-        {{#textarea}}address-textarea{{/textarea}}
-
-        {{#input-submit}}continue{{/input-submit}}
-      {{/form}}
-
-    {{/partials-form}}
-
-  {{/content}}
-
-{{/layout}}
