+++ apps/gro/views/address.html
@@ -13,21 +13,25 @@
     {{> partials-navigation}}
   {{/propositionHeader}}
 
-  {{$header}}{{#t}}pages.post.header{{/t}}{{/header}}
+  {{$header}}{{#t}}pages.postcode.header{{/t}}{{/header}}
 
   {{$content}}
 
     {{<partials-form}}
 
       {{$form}}
-        <p>
-          <span class="postcode">{{values.postcode}}</span>
-          <span class="link"><a href="{{baseUrl}}/postcode">{{#t}}pages.address.edit{{/t}}</a></span>
-        </p>
+        {{#isDomestic}}
+          {{> partials-amend-postcode}}
+          {{#postcodeApiMessageKey}}
+            <div class="alert-info">
+              <span class="info-message">
+                {{#t}}pages.no-postcode.postcode-api.{{postcodeApiMessageKey}}{{/t}}
+              </span>
+            </div>
+          {{/postcodeApiMessageKey}}
+        {{/isDomestic}}
 
-        <p class="address">{{#select}}address{{/select}}</p>
-
-        <p><span class="link"><a href="{{baseUrl}}/address-inside">{{#t}}pages.address.cantfind{{/t}}</a></span></p>
+        {{#textarea}}address-textarea{{/textarea}}
 
         {{#input-submit}}continue{{/input-submit}}
       {{/form}}
