--- apps/gro/views/no-postcode.html
@@ -1,48 +1,0 @@
-{{<layout}}
-  {{$step}}Step 2{{/step}}
-
-  {{$journeyHeader}}
-    {{#t}}journey.header{{/t}}
-  {{/journeyHeader}}
-
-  {{$validationSummary}}
-    {{> partials-validation-summary}}
-  {{/validationSummary}}
-
-  {{$propositionHeader}}
-    {{> partials-navigation}}
-  {{/propositionHeader}}
-
-  {{$header}}{{#t}}pages.postcode.header{{/t}}{{/header}}
-
-  {{$content}}
-
-    {{<partials-form}}
-
-      {{$form}}
-        <p>
-          <span class="postcode">{{values.postcode}}</span>
-            {{#backLink}}
-              <span class="link">
-                <a href="{{#url}}{{backLink}}{{/url}}">{{#t}}pages.address.edit{{/t}}</a>
-              </span>
-            {{/backLink}}
-        </p>
-
-        <div class="alert-info">
-          <span class="info-message">
-            {{#t}}pages.no-postcode.postcode-api.{{postcodeApiMessageKey}}{{/t}}
-          </span>
-        </div>
-
-        {{#textarea}}address-textarea{{/textarea}}
-
-        {{#input-submit}}continue{{/input-submit}}
-
-      {{/form}}
-
-    {{/partials-form}}
-
-  {{/content}}
-
-{{/layout}}
